requirements.txt
# Core dependencies for the Discord bot
discord.py
python-dotenv
requests
google-generativeai
